Renaming view model namespace from SearchFunding to Opportunities. Also putting in place "Detail" action/view. Just imported heroku prototype HTML for now.
