+++ src/SFA.DAS.LevyTransferMatching.Web/Models/Opportunities/DetailViewModel.cs
@@ -1,0 +1,6 @@
+﻿namespace SFA.DAS.LevyTransferMatching.Web.Models.Opportunities
+{
+    public class DetailViewModel
+    {
+    }
+}
