+++ src/SFA.DAS.LevyTransferMatching.Web/Models/Opportunities/IndexViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace SFA.DAS.LevyTransferMatching.Web.Models.SearchFunding
+namespace SFA.DAS.LevyTransferMatching.Web.Models.Opportunities
 {
     public class IndexViewModel
     {
