+++ src/SFA.DAS.LevyTransferMatching.Web/Orchestrators/IOpportunitiesOrchestrator.cs
@@ -1,4 +1,4 @@
-﻿using SFA.DAS.LevyTransferMatching.Web.Models.SearchFunding;
+﻿using SFA.DAS.LevyTransferMatching.Web.Models.Opportunities;
 using System.Threading.Tasks;
 
 namespace SFA.DAS.LevyTransferMatching.Web.Orchestrators
@@ -6,5 +6,6 @@ namespace SFA.DAS.LevyTransferMatching.Web.Orchestrators
     public interface IOpportunitiesOrchestrator
     {
         Task<IndexViewModel> GetIndexViewModel();
+        Task<DetailViewModel> GetDetailViewModel(string encodedId);
     }
 }
