+++ src/SFA.DAS.LevyTransferMatching.Web/Views/Opportunities/Detail.cshtml
@@ -1,0 +1,72 @@
+﻿@model SFA.DAS.LevyTransferMatching.Web.Models.Opportunities.DetailViewModel
+
+@{
+    ViewData["Title"] = "Manage Apprenticeships";
+}
+
+<a href="@Url.Action("index")" class="govuk-back-link">Back</a>
+<main class="govuk-main-wrapper govuk-main-wrapper--auto-spacing" id="main-content" role="main">
+    <div class="govuk-width-container form-entry" id="transfers-hub">
+        <div class="govuk-grid-row">
+            <div class="govuk-grid-column-two-thirds">
+                <span class="govuk-caption-m">Find a business to fund training for your apprentices</span>
+                <h1 class="govuk-heading-xl">Transfer fund details for ####</h1>
+
+                <!-- PLEDGE DETAILS - START -->
+                <h2 class="govuk-heading-m">Details of the opportunity</h2>
+
+                <section class="employer-details side-panel  applications-details">
+                    <p class="govuk-heading-s"><span class="pledge-name-answer">####</span> wants to fund apprenticeship training in:</p>
+                    <p class="details">
+                        Location: <strong class="pledge-location-answer">####</strong><br>
+                        Sector: <strong class="pledge-sector-answer">####</strong><br>
+                        Type of job role: <strong class="pledge-training-answer">####</strong><br>
+                        Level: <strong class="pledge-level-answer">####</strong>
+                    </p>
+                    <div class="total">
+                        <p class="govuk-heading-s">Total funding available: up to £<span class="pledge-value-answer value-format">####</span> available in 2021/22</p>
+                    </div>
+                </section>
+                <!-- PLEDGE DETAILS - END -->
+
+                <form method="post" action="" id="apply-application-form">
+
+                    <div class="govuk-form-group">
+                        <fieldset class="govuk-fieldset">
+                            <legend class="govuk-fieldset__legend">
+                                <h1 class="govuk-fieldset__heading govuk-label--s">Would you like to apply for the transfer funds?</h1>
+                            </legend>
+                            <div class="govuk-radios govuk-radios--conditional" data-module="govuk-radios">
+                                <div class="govuk-radios__item">
+                                    <input class="govuk-radios__input" id="apply-pledge-funds-yes" name="apply-pledge-funds" type="radio" value="Yes" data-aria-controls="conditional-apply-pledge-funds-yes">
+                                    <label class="govuk-label govuk-radios__label" for="apply-pledge-funds-yes">Yes, apply for transfer funds</label>
+                                </div>
+                                <div class="govuk-radios__conditional govuk-radios__conditional--hidden" id="conditional-apply-pledge-funds-yes">
+                                    <div class="govuk-form-group">
+                                        <div class="govuk-warning-text">
+                                            <span class="govuk-warning-text__icon" aria-hidden="true">!</span>
+                                            <strong class="govuk-warning-text__text">
+                                                <span class="govuk-warning-text__assistive">Warning</span>
+                                                You'll need an apprenticeship service account to apply to be considered for funding
+                                            </strong>
+                                        </div>
+                                    </div>
+                                </div>
+                                <div class="govuk-radios__item">
+                                    <input class="govuk-radios__input" id="apply-pledge-funds-no" name="apply-pledge-funds" type="radio" value="No I want to see apprenticeship training providers">
+                                    <label class="govuk-label govuk-radios__label" for="apply-pledge-funds-no">No</label>
+                                </div>
+                            </div>
+                        </fieldset>
+                    </div>
+
+                    <button class="govuk-button" data-module="govuk-button" id="apply-application-continue">Continue</button>
+
+                    <a href="javascript:history.back()" class="govuk-link govuk-link-secondary govuk-link--no-visited-state govuk-back-link">Back</a>
+
+                </form>
+
+            </div>
+        </div>
+    </div>
+</main>
